# projects/control-service/projects/model/apidefs/datajob-api/api.yaml
...
openapi: 3.0.3
servers:
  - url: http://localhost:8092
    description: quickstart-vdk default local setup
info:
  title: Versatile Data Kit Control Service API
  version: '1.0'
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  description: "The Data Jobs API of Versatile Data Kit Control Service.
    Data Jobs allows Data Engineers to implement automated pull ingestion (E in ELT)
    and batch data transformation into a database (T in ELT).
    See also https://github.com/vmware/versatile-data-kit/wiki/Introduction

    The API has resource-oriented URLs, JSON-encoded responses, and uses standard HTTP response codes, authentication, and verbs.
    The API enables creating, deploying, managing and executing Data Jobs in the runtime environment.<br>
    <br>
    ![](https://github.com/vmware/versatile-data-kit/wiki/vdk-data-job-lifecycle-state-diagram.png)
    <br>
    The API reflects the usual Data Job Development lifecycle:<br>
    <li> Create a new data job (webhook to further configure the job, e.g authorize its creation, setup permissions, etc).
    <li> Download keytab. Develop and run the data job locally.
    <li> Deploy the data job in cloud runtime environment to run on a scheduled basis.
    <br><br>
    If Authentication is enabled, pass OAuth2 access token in HTTP header 'Authorization: Bearer [access-token-here]' (https://datatracker.ietf.org/doc/html/rfc6750).
    <br

...
    <br><br>
    <b>API Evolution</b><br>
...
paths:
...
  '/data-jobs/for-team/{team_name}/jobs/{job_name}/executions':
    summary: Execution API is used to query recent executions of Data Jobs and to trigger new executions.
    get:
      tags:
        - Data Jobs Execution
      responses:
        '200':
          $ref: '#/components/responses/ExecutionArray'
        '404':
          $ref: '#/components/responses/Error'
      operationId: dataJobExecutionList
      summary: Get Data Jobs (recent) executions.
      description: <i>(Introduced in v1.0)</i>
    parameters:
      -  name: team_name
         description: Team Name
         schema:
           type: string
         in: path
         required: true
      - name: job_name
        description: Data Job Name.
        schema:
          type: string
        in: path
        required: true
      - name: execution_status
        description: Execution status
        schema:
          type: array
          items:
            type: string
        example: submitted
        in: query
  '/data-jobs/for-team/{team_name}/jobs/{job_name}/deployments/{deployment_id}/executions':
    summary: Execution API is used to query recent executions of Data Jobs and to trigger new executions.
    get:
      tags:
        - Data Jobs Execution
      responses:
        '200':
          $ref: '#/components/responses/ExecutionArray'
        '404':
          $ref: '#/components/responses/Error'
      operationId: dataJobDeploymentExecutionList
      summary: Get Data Jobs (recent) executions.
      description: <i>(Introduced in v1.0)</i>
      parameters:
        - name: execution_status
          description: Execution status
          schema:
            type: array
            items:
              type: string
          example: submitted
          in: query
    post:
      tags:
        - Data Jobs Execution
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataJobExecutionRequest'
      responses:
        '202':
          description: Execution triggered successfully.
          headers:
            Location:
              description: URI of the execution
              schema:
                type: string
        '400':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
      operationId: dataJobExecutionStart
      summary: Trigger Data Job Execution.
      description: |
        <i>(Introduced in v1.0)</i> | Trigger new execution of a Data Job.
        If Data Job deployment is currently being executed it will return an error (409 Conflict).
        Only deployed Data Jobs can be executed. Currently the Data Job arguments ('args') are ignored.
        Note: in the future it will be possible to trigger different deployments.
    parameters:
      -  name: team_name
         description: Team Name
         schema:
           type: string
         in: path
         required: true
      - name: job_name
        description: Data Job Name.
        schema:
          type: string
        in: path
        required: true
      - name: deployment_id
        description: Data job deployment ID. Currently it is being ignored as multiple deployments are not implemented.
        schema:
          $ref: '#/components/schemas/DataJobDeploymentId'
        in: path
        required: true
  '/data-jobs/for-team/{team_name}/jobs/{job_name}/executions/{execution_id}':
    summary: Data Job Execution details.
    get:
      tags:
        - Data Jobs Execution
      responses:
        '200':
          $ref: '#/components/responses/Execution'
        '404':
          $ref: '#/components/responses/Error'
      operationId: dataJobExecutionRead
      summary: Get Data Job Execution details.
      description: <i>(Introduced in v1.0)</i>
    delete:
      tags:
        - Data Jobs Execution
      responses:
        '200':
          description: Canceled execution successfully.
        '404':
          $ref: '#/components/responses/Error'
      operationId: dataJobExecutionCancel
      summary: Cancel (if running) Data Job Execution
      description: <i>(Introduced in v1.0)</i> | Cancel the execution of a Data Job with a given execution ID.
    parameters:
      -  name: team_name
         description: Team Name
         schema:
           type: string
         in: path
         required: true
      - name: job_name
        description: Data Job Name.
        schema:
          type: string
        in: path
        required: true
      - name: execution_id
        description: Execution ID.
        schema:
          type: string
        in: path
        required: true
  '/data-jobs/for-team/{team_name}/jobs/{job_name}/executions/{execution_id}/logs':
    summary: Data Job Execution Logs.
    get:
      tags:
        - Data Jobs Execution
      responses:
        '200':
          $ref: '#/components/responses/ExecutionLogs'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
      operationId: dataJobLogsDownload
      summary: |
          Download data job logs.
          This API is guranteed to provide logs only if the jobs is currently running.
          For logs from older job executions - use logsUrl field passed by GET execution API or jobsQuery API.
      description: <i>(Introduced in v1.2.9)</i>
      parameters:
        -  name: team_name
           description: Team Name
           schema:
             type: string
           in: path
           required: true
        - name: job_name
          description: Data Job Name.
          schema:
            type: string
          in: path
          required: true
        - name: execution_id
          description: Execution ID.
          schema:
            type: string
          in: path
          required: true
        - name: tail_lines
          description: tail_lines when set to positive value, the number of lines from the end of the logs to return.
          schema:
            type: integer
            default: 0
            example: 10
          required: false
          in: query
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      description: Contains description for one or more errors detected.
      required:
        - messages
      type: object
      properties:
        messages:
          description: Error messages
          type: array
          items:
            type: string
          example:
            - Data Job starshot-processing-vmc-fact-daily not found
    DataJobDeployment:
      description: A deployment of the Data Job
      type: object
      properties:
        vdk_version:
          description: A specific VDK version to use
          type: string
          example: 2.1
        job_version:
          description: Job version (can be Git commit)
          type: string
          example: 11a403ba
        python_version:
          description: A python release version (supported by the service) to be used for job deployments.
          type: string
          example: '3.9'
        mode:
          $ref: '#/components/schemas/DataJobMode'
        id:
          $ref: '#/components/schemas/DataJobDeploymentId'
        enabled:
          description: Enable/disable flag
          type: boolean
          example: false
        deployed_by:
          description: User or service that deployed the Data Job
          type: string
          example: auserov@example.mail.com
        deployed_date:
          description: The Data Job deployment date
          type: string
          format: date-time
        schedule:
          $ref: '#/components/schemas/DataJobSchedule'
        resources:
          $ref: '#/components/schemas/DataJobResources'
...
    DataJobExecutionRequest:
      description: Request to start execution of Data Job. The job must have been deployed before that (see Deployment API) and will run its latest version.
      type: object
      properties:
        started_by:
          description: User or service that started the execution (e.g manual/auserov@example.mail.com or scheduled/runtime)
          type: string
          example: schedule/runtime
        args:
          description: Data Job arguments
          type: object
          additionalProperties: true
          example:
            key: value
    DataJobExecution:
      description: Executions of a Data Job
      type: object
      properties:
        id:
          description: Data Job Execution ID
          type: string
          example: starshot-processing-vmc-fact-daily-2018623174356
        job_name:
          description: Data Job name
          type: string
          example: starshot-processing-vmc-fact-daily
        status:
          description: The current status
          enum:
            - submitted
            - running
            - succeeded
            - cancelled
            - skipped
            - user_error
            - platform_error
          type: string
          example: submitted
        type:
          description: Execution type - manual or scheduled
          enum:
            - manual
            - scheduled
          type: string
          example: scheduled
        start_time:
          description: Start of execution
          type: string
          format: date-time
        end_time:
          description: Start of execution
          type: string
          format: date-time
        started_by:
          description: User or service that started the execution (e.g manual/auserov@example.mail.com or scheduled/runtime)
          type: string
          example: schedule/runtime
        logs_url:
          description: |
            URL link to persisted logs in central location. Logs generally should be available for longer time.
            The link is available only if operators have configured it during installation of Control Service.
            During install operators can configure logs to be persisted to log aggregator service whose link can be exposed here.
          type: string
          example: http://logs/jobs?filter=job-name
        message:
          description: Message (usually error) during execution
          type: string
          example:
        op_id:
          description: Operation id used for tracing calls between different services
          type: string
        deployment:
          $ref: '#/components/schemas/DataJobDeployment'
    DataJobExecutionLogs:
      description: Executions of a Data Job
      type: object
      properties:
        logs:
          description: The logs of the data job execution.
          type: string
...
    DataJobMode:
      description: The execution mode that the data job is deployed in. Data Jobs used for development or testing purposes are marked as testing.
        This is used by Operations team on platform rollout and infrastructure changes adoption. For example, rollout and validation of testing jobs first, then proceeding with release data jobs. Also, testing and release jobs may have different limits and SLA targets.
      enum:
        - testing
        - release
      type: string
      example: release
    DataJobDeploymentId:
      description: |
        String that identifies a single deployment of a Data Job.
        Currently only one single deployment per Data Job is possible.<br>
        In the future:<br>
        It's recommended to use following ids - development, testing, production.
        However users are free to come up with their own.
        For example, this enables the creation of 3 different deployments, using the same Data Job code:<br>
        `development  deployment  --deployment-id development`<br>
        `testing deployment  --deployment-id testing`<br>
        `production deployment  --deployment-id prod`
      type: string
      example: release
    DataJobResources:
      description: Resource cofiguration of a data Data Job Deployment.
      type: object
      properties:
        cpu_request:
          format: float
          description: Initial CPU shares in deciCores (1 dCore = 0.1 Core = 100 mCores)
          type: number
          example: 10
        cpu_limit:
          format: float
          description: Max CPU shares in deciCores (1 dCore = 0.1 Core = 100 mCores)
          type: number
          example: 20
        memory_request:
          format: int32
          description: Initial Memory in MiB.
          type: integer
          example: 1024
        memory_limit:
          format: int32
          description: Max Memory in MiB.
          type: integer
          example: 2048
...
    DataJobSchedule:
      description: Schedule configuration
      type: object
      properties:
        schedule_cron:
          example: 0 0 13 * 5
          description: |
            For format see https://en.wikipedia.org/wiki/Cron<br>
            The cron expression is evaluated in UTC time.
            If it is time for a new job run and the previous job run hasn't finished yet, the cron job kills and replaces the currently running job run with a new job run.
            Jobs configured to run more often than once per hour are not supported and their schedule may be overridden by the platform.
            To distribute load evenly, Administrators may override the minute you specified.
            Use https://crontab.guru for help.
          type: string
...
  responses:
...
    Error:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: A response for any errors detected
...
    ExecutionArray:
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/DataJobExecution'
      description: Data Job executions
    Execution:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DataJobExecution'
      description: Execution details
    ExecutionLogs:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DataJobExecutionLogs'
      description: Deployment properties
...
security:
  - bearerAuth: []
...